parse DN correctly in header helper
