+++ sunpy/map/tests/test_header_helper.py
@@ -210,12 +210,12 @@ def test_quantity_input(map_data, hpc_coord):
     header = make_fitswcs_header(map_quantity, hpc_coord, unit=override_unit)
     assert header['bunit'] == override_unit.to_string('fits')
 
-
-def test_unit_as_string(map_data, hpc_coord):
+@pytest.mark.parametrize('unit_string', ['ct / (pix s)', 'DN', 'DN / s'])
+def test_unit_as_string(unit_string, map_data, hpc_coord):
     # Test that unit can be passed in as a string
-    map_unit = u.Unit('ct / (pix s)')
+    map_unit = u.Unit(unit_string)
     header = make_fitswcs_header(map_data, hpc_coord, unit=map_unit.to_string())
-    assert header['bunit'] == map_unit.to_string('fits')
+    assert header['bunit'] == unit_string
 
 
 def test_invalid_inputs(map_data, hcc_coord, hpc_coord_notime, hpc_coord):
